Don't "have" a measurer available until they've identified ...
Because on tityos randos on the Internet connect to the coordinator
thinking it's a torrent client.

This fix doesn't add connections to the list of measurers until they've
identified. It also allows for key errors when a connection dies and we
attempt to remove the connection from the list of identified measurers.