Change coord/measurer to give diff meaning to measurement params ...
Controller gives a relay fp, num of measurers, and num of conns per
measurer. But with recent changes in how we want to think about things
and with the change in tor clients supporting multiple measurement
circuits, it is finally time to differentiate in code the difference
between measurement computers (m-pc) and measurement ph processes (m-proc).

The num-of-measurers param that the controller gives still means total
num of ph measurer processes (m-procs). The coordinator uses it to verify
the right number of m-procs are available.

The num-of-conns-per-measurer param now means the number of conns an
m-pc should make in total spread across its m-procs.